requirements.txt
allure-pytest==2.6.3
allure-python-commons==2.6.3
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
configparser==3.7.4
crayons==0.2.0
glob2==0.6
idna==2.8
Mako==1.0.10
MarkupSafe==1.1.1
more-itertools==7.0.0
parse==1.12.0
parse-type==0.4.2
pluggy==0.9.0
py==1.8.0
PyMySQL==0.9.3
pytest==4.4.1
pytest-base-url==1.4.1
pytest-bdd==3.1.0
pytest-html==1.20.0
pytest-metadata==1.8.0
pytest-selenium==1.16.0
pytest-variables==1.7.1
requests==2.22.0
selenium==3.141.0
six==1.12.0
urllib3==1.25.3
webdriver-manager==1.7
